Reorganized matches/crud and moved the logic from matches/routers to crud (match_delete).
